Allow pra to recover when fetching raises errors
I did this so that pra will not crash when fetching pull requests
fails (e.g. due to http timeout, missing vpn connection, laptop was
closed for 5 minutes, etc). If there is an error during fetching, the
short message will be printed to the screen and the full error with
stack trace will be logged to ~/.pra.errors.log, and pra will keep
running and retry later.