Let pull sources fail without impacting the others
I did this so if a fetch from a pull source fails due to a problem
that doesn't affect the other sources, we can still display the pull
requests that can be fetched successfully. I also changed the error
display to print a generic message telling the user where to look to
find detailed errors.