Pre clear only excess lines on draw
Why you made the change:

I did this to eliminate a the complete clear that was happening for all
lines that was happening before. Instead now the clearing happens on
each line update as it was before, but also only for the extra lines
that no longer should exist. This seems to have resolved the flashing
that was happening on navigation changes as they use the same draw
algorithm as the update for simplicity sake.