Correct logic for clearing old PRs
Why you made the change:

There was a problem when a PR was closed, Pra wasn't clearing the extra
PRs from the list when the list shrank in length.

How the change addresses the need:

Correcting this logic makes it so that the extra cached PRs are now
cleared when the list of PRs shrinks.