Make pagination limit 300
Why you made the change:

I did this because I wanted it to be some outrageous number so that we wouldn't
have issues with pull requests getting clipped by pagination in most cases.
Basically, it is because I am too lazy to deal with implementing pagination on
the client side right now.