Add pull request filtering
I did this because sometimes I am looking for pull requests from a specific
repository, or author, or tag, ect and I'd like to filter them quickly.

This adds a listener for a `/` key press for cursor which gathers all
following key presses as a string until `ENTER` is pressed. Once `ENTER` is
pressed it reduces the current pull requests collection down to a collection
of pull requests that have partial string matches in any of the columns
rendered on the screen. This filtering is slightly smart in that if you use
uppercase letters it will preserve matching by uppercase letters. If you use
lowercase letters, it will match upper and lowercase.

I found that there are some caveats that I can explore if needed. The
`update_at` is a date time that during presentation is being turned into
`.to_time.ago.to_words`. Currently I am not matching against that string
representation so you cannot filter by something in that column e.g. `month`

Also, filtering is based off of what you currently see. For example, I filter
by `approved` and I would see all the pull requests with the word approved in
it. I can then press `/` again and type `ryan` to filter all the approved pull
requests by the word `ryan`. If I want to reset all my filters I'll have to
press `r` to force a refresh to happen and get all of the pull requests.