Filter using presenter strings
I did this so that we can filter the list of pull requests by what the user is
seeing on the screen. It accomplishes this by creating a presenter that
renders the strings for string matching. For example, you can now type in
`month` and see all the pull requests that match `month` in all columns
including `updated_at`. This might be useful for finding long lived pull
requests among many.